Add feature to close ChannelUnbounded (#3)
`ChannelUnbounded` can be "closed" like a socket or a pipe to indicate
to the receiver that no more messages will ever be sent.

Notes about this behavior:

Important!
1. A sender can close a channel to indicate that no more values will be sent.

2. The receiver can continue reading the elements from the channel if the
channel still has elements in it, i. e. if the channel is not yet drained.

3. Only the sender should close a channel, never the receiver.

4. Closing is only necessary when the receiver must be told
there are no more values coming from the sender from that moment.